fix: save epoch as int instead of float

+++ mapillary_tools/uploader.py
@@ -862,7 +862,7 @@ def send_videos_for_processing(video_import_path, user_name, user_email=None, us
             # Blackvue actually reports endtime in the created_at exif field
             endtime = get_video_start_time(video)
             video_start_time = (endtime - datetime.timedelta(seconds=duration))
-            video_start_timestamp = (((video_start_time - datetime.datetime(1970, 1, 1)).total_seconds())-3600)*1000 #TODO Currently adding an hour due to error in camera config. How to fix long term?
+            video_start_timestamp = int((((video_start_time - datetime.datetime(1970, 1, 1)).total_seconds())-3600)*1000) #TODO Currently adding an hour due to error in camera config. How to fix long term?
 
             upload_video_for_processing(
                 video, video_start_timestamp, max_attempts, credentials, user_permission_hash, user_signature_hash, request_params, organization_username, organization_key, private)
